Refactor (#846)
