+++ apps/mobile-app/android/app/src/main/java/net/aliasvault/app/autofill/utils/CredentialMatcher.kt
@@ -1,4 +1,4 @@
-package net.aliasvault.app.autofill
+package net.aliasvault.app.autofill.utils
 
 import net.aliasvault.app.vaultstore.models.Credential
