+++ apps/mobile-app/android/app/src/main/java/net/aliasvault/app/autofill/utils/ImageUtils.kt
@@ -1,4 +1,4 @@
-package net.aliasvault.app.autofill
+package net.aliasvault.app.autofill.utils
 
 import android.graphics.Bitmap
 import android.graphics.BitmapFactory
